Revert to an uncompressed MccTable.
The compressed MccTable didn't have enough space for all the timezones or languages in the world.
We can remove the timezones entirely by looking up the timezone from the ICU data.  Since there
is not much to be gained from the binary format and it had real limitations we chose
to revert it.

Fixed a bug where the wifi channel was never being set.

bug:2284210
Change-Id: If9ca89f7053f937f555430b061c60cd7eec86d8e