Support UI visibility and restarting in download manager.
First, this change adds a column to the downloads table specifying
whether a download should be displayed in the system downloads UI.  It
adds a public method to set this parameter when requesting a download,
and a hidden method to filter queries based on this paramter (such
filtering could be made public later, but it's not strictly necessary
right now).

Second, this change adds support for restarting a completed/failed
download as a hidden method on DownloadManager.  Currently it only
works from the download manager's process - it'll be used by the new
download manager UI.

Change-Id: I15eda1a6e3717d1ce947a810b25ad3540cce809e