Fix a bug in DumpRenderTree2's run_apache2.py script
We use a printf style approach when setting the run command in the
command string to be executed. When doing so, the printf must be
applied to only the relevant part of the string. This avoids problems
in the case that other parts of the string contain special character
sequences, such as '%s' in a custom log directive.

Change-Id: I87df773a2872320386a34cb6a5041a9e0ba6f31e