Fix night/notnight resource folder handling.
The values for NIGHT_NO / NIGHT_YES are shifted later on in the
ResourceTypes header file to match with the ones specified in the
Configuration class file. This is basically the same as it's done for
the SCREEN_LONG_NO / SCREEN_LONG_YES constants.

Since the values are also masked later on by 0x30 (MASK_UI_MODE_NIGHT)
the system currently recognizes resources specified in night/notnight
folders as if they were resources which can be used for any of both
modes. This results in an error saying that the resources are already
specified somewhere else.

Should fix: http://b/issue?id=3022577

Change-Id: I554b37231ea4c69a8625e519150473fddce7d8ed