Pass whether a scroll is user-scrolled or not to webkit.
This is to eliminate false scrollEvent to webkit.  And the side effect
is that the history scroll position will be correctly restored;
otherwise, webkit think there's user scrolled and not send back history
scroll position.

Particulary for NEW_PICTURE_MSG from webkit, the scroll event because of
it are not user-scrolled.

issue: 3063945
Change-Id: I9fdbfcb8baf07baf2331e2c4eed66f88aea744ef