Only AV formats should be handled by MediaPlayer (Java side)
Must be submitted with corresponding C++ change (I50da7599).

When opening a text file in the browser, WebKit was attempting to
display it as a movie. This is because the media player now reports
that it can handle various document formats, including plain text.

See http://b/issue?id=3132917

This CL updates WebViewCore.supportsMimeType() to explicitly check
for audio and video formats only. This is called from WebKit via
JNI, and doesn't seem to be used elsewhere, so it should be a safe
change to make. Renamed to isSupportedMediaMimeType() for clarity.

Test: push a text file to /sdcard/foo.txt, and open it in the
browser (as file:///sdcard/foo.txt).

Change-Id: Id39b770b05ced7c976d8043649845f26115b24c9