Stabilize reflow, aka, the fit-to-screen option in settings.
This change is only for tablet device which has large screen and
good for fixed viewport.

The reflow will only happen in first layout and double tap if the
text wrap scale is not desired.

The initial text wrap scale is set to the reading level, so that if
the web page is in overview mode, after a double tap, the text will
be nicely shown full screen without any reflow.

All the previous reflowing, such as done when zoomed out, are disabled for
tablet device.

issue: 3003933
Change-Id: I8f460704bb27c50d6f5fa259ca2d5007648adfba